+++ dart/jovial_misc/example/example.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'dart:io';
 import 'package:intl/intl.dart';
 import 'package:jovial_misc/io_utils.dart';
+import 'package:jovial_misc/io_utils_native.dart';
 import 'package:jovial_misc/isolate_stream.dart';
 
 ///