+++ dart/jovial_misc/test/io_utils_tests.dart
@@ -12,6 +12,7 @@ import 'package:intl/intl.dart';
 import 'package:pointycastle/export.dart';
 
 import 'package:jovial_misc/io_utils.dart';
+import 'package:jovial_misc/io_utils_native.dart';
 import 'package:jovial_misc/isolate_stream.dart';
 
 const _paranoia = 1;